🛡️ Sentinel: CRITICAL Fix Integer Overflow in Memory Allocations
- Explicitly cast integer operands to `size_t` before multiplying them within `MALLOC` size arguments in `src/fastlp.c` and `src/paralp.c`.
- Prevents 32-bit integer overflow when calculating large dimensions, which previously caused severely undersized heap allocations and buffer overflows.

+++ src/paralp.c
@@ -70,8 +70,8 @@ void paralp(double *obj, double *mat, double *rhs, int *m0 , int *n0, double *op
     }
 
 
-    MALLOC(        a, m*n+m,  double );      
-    MALLOC(       ia, m*n+m,   int );      
+    MALLOC(        a, (size_t)m*n+m,  double );
+    MALLOC(       ia, (size_t)m*n+m,   int );
     MALLOC(       ka, n+m+1,  int );        
     MALLOC(        c, n,   double );
     MALLOC(        b, m,   double );
